Website/Version1.1/requirements.txt
Flask
Flask-WTF
cryptography
Flask-SQLAlchemy
Flask-Login
email-validator
python-dotenv
zxcvbn